LayoutTests/http/tests/site-isolation/mediastream/getUserMedia-audio-starts.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that getUserMedia for audio starts in an OOP iframe.");
jsTestIsAsync = true;

if (window.testRunner) {
    window.testRunner.setUserMediaPermission(true);
}

window.addEventListener("message", function (event) {
    shouldBe("event.data", "'success'");
    finishJSTest();
});
</script>
<iframe allow="microphone *" src="http://localhost:8000/site-isolation/resources/getUserMedia-audio-starts-frame.html"></iframe>
