LayoutTests/http/tests/site-isolation/mediastream/getUserMedia-audio-starts-expected.txt

Tests that getUserMedia for audio starts in an OOP iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is 'success'
PASS successfullyParsed is true

TEST COMPLETE
